+++ tests/unit/FunctionsTest.php
@@ -12,6 +12,9 @@ namespace WikibaseSolutions\CypherDSL\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use TypeError;
+use function WikibaseSolutions\CypherDSL\allShortest;
+use function WikibaseSolutions\CypherDSL\allShortestPaths;
+use function WikibaseSolutions\CypherDSL\anyPath;
 use WikibaseSolutions\CypherDSL\Expressions\Literals\Boolean;
 use WikibaseSolutions\CypherDSL\Expressions\Literals\Float_;
 use WikibaseSolutions\CypherDSL\Expressions\Literals\Integer;
@@ -29,13 +32,22 @@ use function WikibaseSolutions\CypherDSL\list_;
 use function WikibaseSolutions\CypherDSL\literal;
 use function WikibaseSolutions\CypherDSL\map;
 use function WikibaseSolutions\CypherDSL\node;
+use WikibaseSolutions\CypherDSL\Patterns\AllShortest;
+use WikibaseSolutions\CypherDSL\Patterns\AllShortestPaths;
+use WikibaseSolutions\CypherDSL\Patterns\AnyPath;
 use WikibaseSolutions\CypherDSL\Patterns\Direction;
 use WikibaseSolutions\CypherDSL\Patterns\Node;
 use WikibaseSolutions\CypherDSL\Patterns\Relationship;
+use WikibaseSolutions\CypherDSL\Patterns\Shortest;
+use WikibaseSolutions\CypherDSL\Patterns\ShortestGroups;
+use WikibaseSolutions\CypherDSL\Patterns\ShortestPath;
 use WikibaseSolutions\CypherDSL\Query;
 use function WikibaseSolutions\CypherDSL\query;
 use function WikibaseSolutions\CypherDSL\raw;
 use function WikibaseSolutions\CypherDSL\relationship;
+use function WikibaseSolutions\CypherDSL\shortest;
+use function WikibaseSolutions\CypherDSL\shortestGroups;
+use function WikibaseSolutions\CypherDSL\shortestPath;
 use function WikibaseSolutions\CypherDSL\string;
 use function WikibaseSolutions\CypherDSL\variable;
 
@@ -246,5 +258,77 @@ final class FunctionsTest extends TestCase
 
         // @phpstan-ignore-next-line
         raw([]);
+    }
+
+    public function testShortestPath(): void
+    {
+        $path = node()->relationshipTo(node());
+        $shortestPath = shortestPath($path);
+
+        $this->assertInstanceOf(ShortestPath::class, $shortestPath);
+        $this->assertSame('shortestPath(()-->())', $shortestPath->toQuery());
+    }
+
+    public function testAllShortestPaths(): void
+    {
+        $path = node()->relationshipTo(node());
+        $allShortestPaths = allShortestPaths($path);
+
+        $this->assertInstanceOf(AllShortestPaths::class, $allShortestPaths);
+        $this->assertSame('allShortestPaths(()-->())', $allShortestPaths->toQuery());
+    }
+
+    public function testShortest(): void
+    {
+        $path = node()->relationshipTo(node());
+        $shortest = shortest($path, 2);
+
+        $this->assertInstanceOf(Shortest::class, $shortest);
+        $this->assertSame('SHORTEST 2 (()-->())', $shortest->toQuery());
+    }
+
+    public function testShortestIntegerType(): void
+    {
+        $path = node()->relationshipTo(node());
+        $shortest = shortest($path, new Integer(2));
+
+        $this->assertInstanceOf(Shortest::class, $shortest);
+        $this->assertSame('SHORTEST 2 (()-->())', $shortest->toQuery());
+    }
+
+    public function testAllShortest(): void
+    {
+        $path = node()->relationshipTo(node());
+        $allShortest = allShortest($path);
+
+        $this->assertInstanceOf(AllShortest::class, $allShortest);
+        $this->assertSame('ALL SHORTEST (()-->())', $allShortest->toQuery());
+    }
+
+    public function testShortestGroups(): void
+    {
+        $path = node()->relationshipTo(node());
+        $shortestGroups = shortestGroups($path, 3);
+
+        $this->assertInstanceOf(ShortestGroups::class, $shortestGroups);
+        $this->assertSame('SHORTEST 3 GROUPS (()-->())', $shortestGroups->toQuery());
+    }
+
+    public function testShortestGroupsIntegerType(): void
+    {
+        $path = node()->relationshipTo(node());
+        $shortestGroups = shortestGroups($path, new Integer(3));
+
+        $this->assertInstanceOf(ShortestGroups::class, $shortestGroups);
+        $this->assertSame('SHORTEST 3 GROUPS (()-->())', $shortestGroups->toQuery());
+    }
+
+    public function testAnyPath(): void
+    {
+        $path = node()->relationshipTo(node());
+        $anyPath = anyPath($path);
+
+        $this->assertInstanceOf(AnyPath::class, $anyPath);
+        $this->assertSame('ANY (()-->())', $anyPath->toQuery());
     }
 }